Change product page to smaller components
