+++ src/components/payment/delivery.tsx
@@ -1,0 +1,30 @@
+import { CSSProperties, FC } from "react"
+
+
+interface Props {
+
+}
+
+
+const DeliveryPage: FC<Props> = (props) => {
+
+
+
+
+    return (
+        <>
+            
+            
+     
+        </>
+    )
+}
+
+
+
+
+
+
+
+
+export default DeliveryPage
