+++ src/components/product/compareProduct.tsx
@@ -9,10 +9,12 @@ import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
 import { products, } from '../../data/products';
 import { colors } from '../../data/color';
-import { CSSProperties } from 'react';
+import { CSSProperties, FC } from 'react';
+import CartButton from '../interaction/cartButton';
+import ProductButton from '../interaction/toProductButton';
 
 interface Props {
-  
+
 }
 
 
@@ -32,7 +34,7 @@ const StyledTableCell = styled(TableCell)(({ theme }) => ({
     color: theme.palette.common.white,
     fontSize: 24,
   },
-  
+
 }));
 
 const StyledTableRow = styled(TableRow)(({ theme }) => ({
@@ -82,19 +84,19 @@ function createData(
   productBas: string,
   productStandard: string,
   productPlus: string,
-  
+
 ) {
   return { name, productBas, productStandard, productPlus };
 }
 
 const rows = [
-  
+
   createData("Bokföring", "", "", ""),
   createData("Leverantörsregister", "✅", "✅", "✅"),
   createData("Leverantörsfakturor", "Obegränsat", "Obegränsat", "Obegränsat"),
   createData("Olika Rapporter", "✅", "✅", "✅"),
   createData("Tolkning av dokument", "från 4,90 kr/st", "från 4,90 kr/st", "från 4,90 kr/st"),
-  
+
 
   createData("Intregration", "", "", ""),
   createData("Fri support", "Obegränsat", "Obegränsat", "Obegränsat"),
@@ -109,14 +111,14 @@ const rows = [
   createData("kundregister", "❌", "✅", "✅"),
   createData("Rapporter", "❌", "✅", "✅"),
   createData("Prislistor", "❌", "✅", "✅"),
- 
+
   createData("Lön", "", "", ""),
   createData("Personalregister", "❌", "❌", "✅"),
   createData("Digitala lönebesked", "❌", "❌", "12 kr/st"),
   createData("När- och frånvaroregistrering", "❌", "❌", "✅"),
   createData("Semesterhantering", "❌", "❌", "✅"),
   createData("Statistikrapportering", "❌", "❌", "✅"),
-  
+
 ];
 
 /* const includes = products.including.map((include) => {
@@ -124,24 +126,27 @@ const rows = [
 })  */
 
 const headerTableCell = products.map((product) => {
-  return  <StyledTableCell key={product.name} align="center" >{product.name}</StyledTableCell>
+  return <StyledTableCell key={product.name} align="center" >{product.name}</StyledTableCell>
 })
 
-const footerTableCell = products.map((product) => {
+const footerTableCell = products.map((products) => {
   return (
-    <StyledTableCell key={product.name} align="center">
-    {product.name}
+    <StyledTableCell key={products.name} align="center">
+      {products.name}
+      <CartButton product={products} />
+      <ProductButton product={products}/>
     </StyledTableCell>
   )
 })
 
 
 
-export default function CustomizedTables() {
+const ProductsCompare: FC<Props> = (props) => {
   return (
-    
-    <TableContainer component={Paper}  style={{display:"flex",justifyContent:"center" }}>
-      <Table sx={{ minWidth: 700, maxWidth: 1000 }} aria-label="customized table">
+
+    <TableContainer component={Paper} style={{ display: "flex", justifyContent: "center", padding: "100px 0px" }}>
+
+      <Table sx={{ minWidth: 700, maxWidth: 1000, boxShadow: 10 }} aria-label="customized table">
         <TableHead>
           <TableRow>
             <StyledTableCell></StyledTableCell>
@@ -151,12 +156,12 @@ export default function CustomizedTables() {
         <TableBody>
           {rows.map((row) => (
             <StyledTableRow key={row.name}>
-              <StyledTableCell  component="th" scope="row" style={{...firstColumn}}>
+              <StyledTableCell component="th" scope="row" style={{ ...firstColumn }}>
                 {row.name}
               </StyledTableCell>
-              <StyledTableCell align="center" style={{...secondColumn}}>{row.productBas}</StyledTableCell>
-              <StyledTableCell align="center" style={{...thirdColumn}}>{row.productStandard}</StyledTableCell>
-              <StyledTableCell align="center" style={{...fourthColumn}}>{row.productPlus}</StyledTableCell>
+              <StyledTableCell align="center" style={{ ...secondColumn }}>{row.productBas}</StyledTableCell>
+              <StyledTableCell align="center" style={{ ...thirdColumn }}>{row.productStandard}</StyledTableCell>
+              <StyledTableCell align="center" style={{ ...fourthColumn }}>{row.productPlus}</StyledTableCell>
             </StyledTableRow>
           ))}
         </TableBody>
@@ -174,7 +179,7 @@ export default function CustomizedTables() {
 const firstColumn: CSSProperties = {
   backgroundColor: "white",
   color: "black",
-  fontWeight: "bold", 
+  fontWeight: "bold",
   maxWidth: "300px",
   fontSize: 17,
 }
@@ -193,3 +198,5 @@ const fourthColumn: CSSProperties = {
   color: "white",
   fontSize: 17,
 }
+
+export default ProductsCompare
